move RNN module to separate dir
